IN : Move migrate into PrestShopBundle namespace
